Finihsed off test plans and updated version.
