+++ RoundTheCode.ByteTurn/RoundTheCode.ByteTurn.nuspec
@@ -3,7 +3,7 @@
 	<metadata>
 		<id>RoundTheCode.ByteTurn</id>
     <title>Byte Turn</title>
-		<version>0.1.0.0</version>
+		<version>1.0.0.0</version>
 		<authors>Round The Code</authors>
 		<description>File manager tool for a .NET application that allows retrieving the properties of an individual file or directory, such as the name, the size, the extension name, when it was created and much more. A directory listing can also be retrieved listing all the properties for the files and directories that reside within that path. The tool also allows the ability to create, move, copy and delete files and directories.
